move constants to toplevel package
